requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
pymysql==1.1.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2